--- a/schema.sql
+++ b/schema.sql
@@ -49,3 +49,5 @@ CREATE TABLE visits(
 
 -- Add an email column to your owners table
 ALTER TABLE owners ADD COLUMN email VARCHAR(120);
+
+CREATE INDEX index_animal_id ON visits(animal_id asc);