--- a/schema.sql
+++ b/schema.sql
@@ -51,3 +51,5 @@ CREATE TABLE visits(
 ALTER TABLE owners ADD COLUMN email VARCHAR(120);
 
 CREATE INDEX index_animal_id ON visits(animal_id asc);
+
+CREATE INDEX index_email ON owners(email asc);