In this PR the following changes were implented:

- The following queries are taking too much time (1 sec = 1000ms can be considered as too much time for database query). Try them on your machine to confirm it:

```
SELECT COUNT(*) FROM visits where animal_id = 4;
SELECT * FROM visits where vet_id = 2;
SELECT * FROM owners where email = 'owner_18327@mail.com';
```

- Use `EXPLAIN ANALYZE` on the previous queries to check what is happening. Take screenshots of them.

- While you are making changes, check if they help by running `EXPLAIN ANALYZE`. 

The following images are the results of the queries before and after the improvements:
Before:
<img width="1214" alt="Before Analize select count from visits where animal id is 4" src="https://github.com/George7h/Vet-Clinic-Database/assets/101390796/63186d4b-7ccb-4915-b827-4abd7b32882e">
After:
<img width="1042" alt="After Analize select count from visits where animal id is 4" src="https://github.com/George7h/Vet-Clinic-Database/assets/101390796/61aaffdb-c33a-492c-be96-851575f761df">

Result improved by:  1 204ms

Before:
<img width="1166" alt="Before analize select from visits where vet id is 2" src="https://github.com/George7h/Vet-Clinic-Database/assets/101390796/641ab6fe-7872-4303-bbb6-69e2190f3590">
After:
<img width="953" alt="After analize select from visits where vet id is 2" src="https://github.com/George7h/Vet-Clinic-Database/assets/101390796/e144465e-26c0-42b1-be9f-0af69fce81d9">

Result improved by:  1310ms

Before:
<img width="1218" alt="Before analize select from owners where email is owner_18327@mail compng" src="https://github.com/George7h/Vet-Clinic-Database/assets/101390796/4dc7aa76-5978-4ca7-9a98-13c78f9ddd16">
After:
<img width="955" alt="After analize select from owners where email is owner_18327@mail compng" src="https://github.com/George7h/Vet-Clinic-Database/assets/101390796/8caf891d-6af6-4d3f-a1ad-69bee184c9b5">

Result improved by: 1 234ms